Cleaned up the functions.
The main and mockimg() functions were too messy!
I made a separte source code file: ui.c (ui.h) and
put all the functions to set the default values and
read user input values in that.

Also a structure was defined to keep all the mockimg()
input parameters in it to make the job of going between
these functions easier.