New basis for pix.h
The graph based approach before was very complex,
confusing and hard to work with. But now it all goes
into a simple size_t structure! It is now extremely
easier to work with than before. Also for the 50
default mock galaxies, the speed increased by 20%!
I have fully explained it in the manual (Graph canvas).