Indexing infrastructure in place.
Due to the index, I also had to change
'correctindextop.c', because an [index] link
was added to the top of the page.

Added makerunrequirements.sh so making the two files
to correct the HTML output of the documentation is
now easier.

Added a '.gitignore' so we don't commit any
produced files by mistake.