More accurate central pixel value
Until now there was only integration, but as I have
explained in section 3.1.5, integration is not accurate
in the immediate vicinity of the profile center for
sharp profiles. Therefore I added the option to fist
take the average of a large number of pixels and then
once the tolerance is reached, use integration.

Manual is also fully updated with the new method.