Save PSF as ASCII .conv file.
This is mainly to feed into SExtractor. With this
option if the user requests to view the PSF that
MockGals has made, both a .fits file and a .txt file
will be made. The .txt file, has a CONV NONORM as
the first line which is what SExtractor needs to read
the file.

For this option, I had to change printfarray() in
arraymanip.c and all the functions that called
it were also changed. I added an argument to specify
the format of the output, where for example you can
set 'f' to print float or 'e' to print in exponential.