Angles based on FITS's NAXIS0.
Until now, angles were based on C's first axis,
which was the height of the image (as viewed in ds9)
and increased upwards. But after a tiny correction,
it is now in relation to the width, increasing to
the right (as viewed in ds9).

I also changed the range of random position angles,
it was 0 to 360, but that is too wide, because
the profiles are symmetrical. So I set the range
from -90 to +90.