Two bugs fixed:
In writing the new convolve.c, I had forgot that
I had to devided the final result by the inverse
of the number of pixels! That is why all the objects
had such large values.

By mistake, some time ago, I had put the
findstartingpixel() function after deciding which
profile is needed. So stars would not be positioned
properly! I corrected that!