Added option, PSF made like galaxies.
Until now, I had forgot to actually put an option
to set the tolerance level when a more accurate method
should not be used any more. I done that in this commit.

Until now, the PSF was made with an archaic function I
had written in the stone ages! So I thought its time to
make it use exactly the same function and parameters as
the galaxies.

I also added steps in makeprofile() to calculate the
total Gaussian and total Moffat light (integrated from
zero to infinity). This is just to keep conformity,
because in making the PSF, I will not rely on such
crude methods integration resutls and use the real sum.

I also changed the default PSF fwhm to 3 (was 5).