Verbose mode edited.
Previously in verbose mode, all parameters (even those
that were not used by a particular profile) were printed
in verbose mode. But now for each kind of profile, the
parameters of that profile are printed.

I did some checks using fixed functions for finding the
profiles on random points, since I thought it might
increase the speed. But it was not successful, there
was no speed boost. I guess GCC does the job internally.
It is better like this because it is a generic function
for all different profiles.