Date automatically inserted into manual.
Using a "Here script" and the 'date' command
in run.sh, the date the manual is made is inserted
into the manual automatically. So I don't have to
worry about inserting it! Only the version and edition
numbers have to be changed by hand.

I also made some slight changes to the copyright notice.

generalmacros.texi is now a product of the run.sh script
so there is no need to track it in git any more. So it was
deleted and git is asked to ignore it if it is present.

A comment in mock.c was also corrected which stated
the wrong number of random points used.